Update README.md
India, with a rich heritage of floral diversity, is well-known for its medicinal plant wealth. However, the misidentification of medicinal plants and raw materials is a significant challenge in Ayurvedic Pharmaceutics. Due to similar morphological characteristics, seasonal/geographical variations, and lack of awareness among traders and collectors, several crude drugs are wrongly labeled or substituted in the supply chain. This leads to adulteration, reduced medicinal efficacy, and a loss of trust in Ayurveda.

To solve this, an AI-powered software application using image processing and machine learning algorithms can be developed. The system will allow wholesalers, distributors, and researchers to accurately identify medicinal plants and raw materials by scanning them via a mobile/web application. This will enhance the authenticity of herbal medicines, prevent fraud, and ensure quality control in the supply chain.